function isTruthy
I would use an early return in case of a truthy value. And for all falsy I would use === to compare value and type because 0, false and '' are going to meet each other if conditions resulting in multiple console logs that are not under the spirit of the workshop directions. As it is this would be the result:

true
The boolean value false is falsy
The number 0 is falsy (the only falsy number)
The empty string is falsy (the only falsy string)
The null value is falsy
undefined is falsy
The null value is falsy
undefined is falsy
The boolean value false is falsy
The number 0 is falsy (the only falsy number)
The empty string is falsy (the only falsy string)
The boolean value false is falsy
The number 0 is falsy (the only falsy number)
The empty string is falsy (the only falsy string)

vs the directions:

true
The boolean value false is falsy
The null value is falsy
undefined is falsy
The number 0 is falsy (the only falsy number)
The empty string is falsy (the only falsy string)